support for triggering actions and debug commands
This adds a helper function that makes it easier to trigger
functionality that intellij has keystrokes for. It also adds most of the
debugger commands. The main one missing is SHOW_HOVER, which isn't
really necessary because intellij shoes everything in the UI, so the
inspect command isn't really necessary. There is something similar to
show hover, but I think the functionality is buried and I had trouble
finding the action for it.